small changes do dialog, big updates in main app
dialog is not more or less centered and positioned so that it doesn't cover the data structure presented. In main app, now there is a possibility of scrolling through operations using arrow keys and moreover, data structures appear and update on the screen when choosing an option! Also, resizingCanvas were added for potential use in the future