New 'animated' folder for animated data structures, small updates in main app
After some thought I made a decision to keep animated data structures in a new folder. Main reason is that I want to be able to have access to 'pure' data structures without any noise in the shape of all the animation functions. All new, animated data structures will appear in the animated folder. Apart from that, app.py was updated to take that into account and now all the 'showing' of data structures is performed in the data structure class, which provides more flexibility. Moreover, app.py need not be updated to add any new data structure - they are added automatically.