Animated classes now inherit from original DSs
Instead of copying all the code to new files, the classes for animation (with prefix 'anim_' are now children of the according data structure. __init__.py file for structures has been updated to take that into account as all of those files are now kept in the same folder. New animations for dynamicarray have been added (insert, insertLast)