Remove dynamic loaders from app.py in preference for more explicit operations.
Using inspect to dynamically load objects (eg: classes and functions) can have unintended side effects.

To counter this, I have added an `animations` class attribute to the animation classes which contain functions that are allowed to be added to be assigned to `public_functions` and added
an `AVAILABLE_ANIMATIONS` contstant to the module which is used to populate the `dsmenu` command.